pyproject.toml
[project]
name = "allways"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-unfold (>=0.55.3,<0.56.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"

